+++ cli/bin/supadart.dart
@@ -1,9 +1,8 @@
 import 'dart:io';
 import 'package:args/args.dart';
 import 'package:supadart/config_init.dart';
+import 'package:supadart/new/index.dart';
 import 'package:yaml/yaml.dart';
-import 'package:supadart/generator/generator.dart';
-import 'package:supadart/generator/swagger.dart';
 
 const String version = 'v1.5.4';
 const String red = '\x1B[31m'; // Red text
@@ -112,9 +111,8 @@ void main(List<String> arguments) async {
     print('Failed to fetch database');
     exit(1);
   }
+  final files = supadartRun(databaseSwagger, isDart, isSeparated, mappings);
 
-  final files =
-      generateModelFiles(databaseSwagger, isDart, isSeparated, mappings);
   await generateAndFormatFiles(files, output);
 
   print('$green🎉 Done! $reset');
