+++ cli/lib/new/class/converters.dart
