+++ cli/lib/new/class/to_json.dart
@@ -1,4 +1,4 @@
-import 'swagger.dart';
+import '../swagger/column.dart';
 
 String generateToJsonMethod(
   String className,
