+++ cli/lib/new/standalone/exports.dart
@@ -1,9 +1,8 @@
-import 'utils.dart';
 import 'package:yaml/yaml.dart';
+import '../utils/string_formatters.dart';
+import '../swagger/swagger.dart';
 
-import 'swagger.dart';
-
-String generateModels(DatabaseSwagger swagger, YamlMap? mappings) {
+String generateExports(DatabaseSwagger swagger, YamlMap? mappings) {
   String code = 'library models;\n';
   code += "export 'client_extension.dart';\n";
   code += "export 'supadart_abstract_class.dart';\n";
