+++ cli/lib/new/swagger/utils.dart
@@ -1,3 +1,5 @@
+
+
 class DartType {
   final String type;
   DartType(this.type);
